Adding skeleton for bow_nltk.py (#2)
* Adding skeleton for bow_nltk

* Need to save my work. doesn't have TF-IDF yet which is the point so dont merge yet pls

* TF-IDF not working

* IDK

* Potential TF-IDF calculation

Looks like TF-IDF is being calculated now. NLTK also has a TF-IDF
implementation that I put in for reference. The two are different,
but it looks like we might be now on the right track. Next thing to
do is to implement a guassian model

* Self-made TF-IDF implementation working

Just need to add Guassian prediction now

* Created manual testing harness, and updated manual to include all toggle-able options. it still looks like dog water, though. but it works. I also made a sheets doc with the test output and shared it with you.

* Fixed bow_comparison.py

Co-authored-by: Rob Schmidt <rwschmidt26@gmail.com>
Co-authored-by: Rob Schmidt <rschmi2@uic.edu>